@@ -22,10 +22,7 @@ module Dummy
     #config.time_zone = 'UTC'
     #config.time_zone = 'Africa/Asmara'
     #config.time_zone = 'Central Time (US & Canada)'
-    # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
-    
+
     config.assets.paths << Rails.root.join('vendor', 'assets', 'components')
   end
 end
-
