@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -14,9 +13,9 @@
 ActiveRecord::Schema.define(version: 20150421155740) do
 
   create_table "events", force: :cascade do |t|
-    t.date     "date"
+    t.date "date"
     t.datetime "datetime"
-    t.time     "time"
+    t.time "time"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
